requirements.txt
# Core Web / API
fastapi
uvicorn[standard]

# CLI & UX
click
rich

# AWS
boto3

# GCP
google-cloud-bigquery
google-cloud-compute
google-auth
google-api-core

# Azure
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-costmanagement

# Utilities
python-dotenv
pydantic
httpx
tenacity
apscheduler>=3.10.0

# Storage
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
greenlet

# Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Cache
redis[hiredis]>=5.0.0
